rebasing and checking if correct int values were used in place of HighAvailabilityModeFull and HighAvailabilityModeNone

+++ internal/cluster/validations/validations.go
@@ -703,7 +703,7 @@ func ValidateDiskEncryptionParams(diskEncryptionParams *models.DiskEncryption, D
 }
 
 func ValidateControlPlaneCountWithPlatform(controlPlaneCount *int64, platform *models.Platform) error {
-	if swag.Int64Value(controlPlaneCount) == 1 {
+	if swag.Int64Value(controlPlaneCount) == int64(1) {
 		if platform != nil && platform.Type != nil && *platform.Type != models.PlatformTypeNone && !common.IsPlatformExternal(platform) {
 			return errors.Errorf("Single node cluster is not supported alongside %s platform", *platform.Type)
 		}
